requirements.txt
aiodns>=2.0.0
aiofiles>=0.5.0
aiohttp>=3.6.2
async-timeout>=3.0.1
attrs>=19.3.0
bcrypt>=3.1.7
beautifulsoup4>=4.9.1
bs4>=0.0.1
certifi>=2020.4.5.1
cffi>=1.14.0
chardet>=3.0.4
cryptg>=0.2.post1
cryptography>=2.9.2
geojson>=2.5.0
googletrans>=2.4.0
idna>=2.9
maxminddb>=1.5.4
maxminddb-geolite2>=2018.703
multidict>=4.7.6
numpy>=1.18.4
opencv-python>=4.2.0.34
paramiko>=2.7.1
Pillow>=7.1.2
pyaes>=1.6.1
pyasn1>=0.4.8
pycares>=3.1.1
pycparser>=2.20
PyNaCl>=1.3.0
pyowm>=2.10.0
pysftp==0.2.8
requests>=2.23.0
rsa>=4.0
scipy>=1.4.1
six>=1.14.0
soupsieve>=2.0.1
Telethon==1.13.0
termcolor>=1.1.0
urllib3>=1.25.9
vk-api>=11.8.0
yarl>=1.4.2
youtube-dl>=2020.5.8
hachoir>=3.1.1
